Code rearrangement/cleanup
