+++ src/main/scala/org/fs/mael/ui/components/proxy/ProxyGlobalFieldEditor.scala
@@ -14,7 +14,7 @@ import org.eclipse.swt.widgets._
 import org.fs.mael.core.config.ConfigSetting
 import org.fs.mael.core.config.ConfigSetting._
 import org.fs.mael.core.config.IConfigStore
-import org.fs.mael.core.config.proxy.Proxy
+import org.fs.mael.core.proxy.Proxy
 import org.fs.mael.core.utils.CoreUtils._
 import org.fs.mael.ui.components.ConfigAware
 import org.fs.mael.ui.utils.SwtUtils._
