+++ src/main/scala/org/fs/mael/ui/config/GlobalSettings.scala
@@ -6,7 +6,7 @@ import org.eclipse.jface.preference.IntegerFieldEditor
 import org.fs.mael.core.config.ConfigSetting
 import org.fs.mael.core.config.ConfigSetting._
 import org.fs.mael.core.config.IGlobalConfigStore
-import org.fs.mael.core.config.proxy.Proxy
+import org.fs.mael.core.proxy.Proxy
 import org.fs.mael.core.utils.CoreUtils._
 import org.fs.mael.ui.components.proxy.ProxyGlobalFieldEditor
